requirements.txt
# Add Python dependencies here
langgraph>=0.0.45
langchain>=0.3.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pypdf>=3.17.0
chromadb>=0.5.0
sentence-transformers>=2.2.2
langchain>=0.3.0
langchain-community>=0.0.30
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.2
chromadb>=0.5.0
pypdf>=3.17.0
requests>=2.31.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
